requirements.txt
fastapi
uvicorn
anthropic
streamlit
python-dotenv